Drop all relay dependent components and restructurize code
