+++ saleor/static/js/components/footer.js
@@ -1,0 +1,6 @@
+export default $(document).ready((e) => {
+  let navbarHeight = $('.navbar').outerHeight(true);
+  let footerHeight = $('.footer').outerHeight(true);
+  let windowHeight = $(window).height();
+  $('.maincontent').css('min-height', windowHeight - navbarHeight - footerHeight);
+});
