+++ saleor/static/js/components/variant-picker.js
@@ -5,6 +5,8 @@ import VariantPicker from './variantPicker/VariantPicker';
 import VariantPrice from './variantPicker/VariantPrice';
 import variantPickerStore from '../stores/variantPicker';
 
+import {onAddToCartSuccess, onAddToCartError} from "./cart";
+
 export default $(document).ready((e) => {
   const variantPickerContainer = document.getElementById('variant-picker');
   const variantPriceContainer = document.getElementById('variant-price-component');
